requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AWS Services
boto3==1.34.34
aioboto3==12.3.0
aws-lambda-powertools==2.31.0

# Database & Cache
redis==5.0.1
opensearch-py==2.4.2

# Data Processing
pandas==2.2.0
pyarrow==15.0.0
sqlglot==20.11.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
httpx==0.26.0

# Monitoring & Logging
structlog==24.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
